feature customdialog
created a new customdialog which can have title, description and 2 buttons
if text area for description or title is too small, then the area will increase in width.
the dialog can be opened by pressing the options key and closed by pressing the replay key.
title, text and buttons are added through interface fields in contentlistscene.